HTLogger Documentation

This is a package for logging temperature and humidity data from a compatible device.

Installation

From the Julia REPL:

julia>]
pkg> add https://github.com/MLackner/HTLogger

Usage

From the Julia REPL:

julia> using THLogger
julia> THLogger.run()

This will automatically search for correct device and start logging in a log/ folder that will be created in the current directory if it doesn't exist yet.

For more customization options refer to the documentation of THLogger.run.

Compatible Devices

The communication happens via the serial port. The device has to implement the following communication patterns:

\I\n should return the logger's identity as thlogger\r\n

\T\n should return the temperature as for example 23.21\r\n

\H\n should return the humidity as for example 50.21\r\n

The code in >this repository< runs on an Arduino Uno with an HYT939 Temperature/Humidity sensor.

Functions

HTLogger.run — Method

run(;path="log", baudrate=9600, port="", interval=5, lines_per_file=135_000, debug=false)

Starts logging the temperature and humidity data from the connected device.

Keyword Arguments

path: the path where the .log files are stored. If the path does not exist, it is created.

baudrate: the baudrate the device is configured to

port: the serial port the device is connected to. If port is an empty string, the program will try to find the device automatically.

interval: the time between the individual measurements in seconds. It is implemented in such a way that the main measurement loop is idle for the interval time. This means that the individual measurements are more than interval seconds apart because additional time is spent reading the data from the device and writing to the file.

lines_per_file=135_000: number of lines in a log file until a new file is created. 1MB roughly corresponds to 28,000 entries.

debug: prints messages that help to debug in case there are problems

Index

HTLogger.run